fix(lint): split routine_storage test family under the 500-line gate
